cmd, core, ethdb: enable Pebble on 32 bits and OpenBSD too
